Moved policy methods from helper class into PolicyHelper
