+++ WDAC-Policy-Wizard/app/src/CustomRuleConditionsPanel.cs
@@ -2413,7 +2413,7 @@ namespace WDAC_Wizard
             // If the policy doesn't support UMCI, prompt the user and set it
             if(this.checkBox_userMode.Checked)
             {
-                if(!Helper.PolicyHasRule(this.Policy.PolicyRuleOptions, OptionType.EnabledUMCI))
+                if(!PolicyHelper.PolicyHasRule(this.Policy.PolicyRuleOptions, OptionType.EnabledUMCI))
                 {
                     DialogResult res = MessageBox.Show("Your policy does not have User mode code integrity (UMCI) enabled so this UMCI rule will not be enforced. Would you like the Wizard to enable UMCI?",
                                                         "Proceed with UMCI Rule Creation?",
@@ -2479,7 +2479,7 @@ namespace WDAC_Wizard
         private void OnLoad(object sender, EventArgs e)
         {
             // If the policy does not support UMCI, uncheck umci and check kmci
-            if(!Helper.PolicyHasRule(this.Policy.PolicyRuleOptions, OptionType.EnabledUMCI))
+            if(!PolicyHelper.PolicyHasRule(this.Policy.PolicyRuleOptions, OptionType.EnabledUMCI))
             {
                 this.checkBox_kernelMode.Checked = true;
                 this.checkBox_userMode.Checked = false;
